refactored the PageAdapters to use shared code from parent
