+++ lib/src/main/java/com/github/omadahealth/slidepager/lib/adapter/SlidePagerAdapter.java
@@ -25,11 +25,7 @@ package com.github.omadahealth.slidepager.lib.adapter;
 
 import android.content.Context;
 import android.content.res.TypedArray;
-import android.support.v4.view.ViewPager;
-import android.view.View;
-import android.view.ViewGroup;
 
-import com.github.omadahealth.slidepager.lib.SlidePager;
 import com.github.omadahealth.slidepager.lib.interfaces.OnSlideListener;
 import com.github.omadahealth.slidepager.lib.interfaces.OnSlidePageChangeListener;
 import com.github.omadahealth.slidepager.lib.utils.Utilities;
@@ -41,144 +37,18 @@ import java.util.Date;
  * Created by stoyan on 4/3/15.
  */
 public class SlidePagerAdapter extends AbstractSlidePagerAdapter<SlideView> {
-    /**
-     * The maximum number of weeks before we change the {@link #getLeftText(int, int)}
-     */
-    private static final int DEFAULT_MAX_WEEKS = 10;
-    /**
-     * The start date of the {@link SlidePager}
-     */
-    private Date mStartDate;
-
-    /**
-     * The end date of the {@link SlidePager}
-     */
-    private Date mEndDate;
-
-    /**
-     * The context of the app
-     */
-    private Context mContext;
-
-    /**
-     * The list of {@link View} used to retain inflated views
-     */
-    private SlideView[] mViews;
-
-    /**
-     * Weeks that this view represents
-     */
-    private int mWeeks;
-
-    /**
-     * The weeks in the program
-     */
-    private final int mMaxWeeks;
-
-    /**
-     * A user defined {@link ViewPager.OnPageChangeListener}
-     */
-    private OnSlidePageChangeListener mUserPageListener;
-
-    /**
-     * The attribute set from xml
-     */
-    private TypedArray mAttributeSet;
-
-    public void setAttributeSet(TypedArray attributeSet) {
-        this.mAttributeSet = attributeSet;
-    }
 
 
-    /**
-     * Calls {@link #SlidePagerAdapter(Context, Date, Date)} with the end date being set
-     * to today
-     *
-     * @param context   The context
-     * @param startDate The start {@link Date} for computing the number of weeks
-     */
     public SlidePagerAdapter(Context context, Date startDate) {
-        this(context, startDate, new Date());
+        super(context, startDate);
     }
 
-    /**
-     * Public constructors to the {@link SlidePagerAdapter}
-     *
-     * @param context   The context
-     * @param startDate The start {@link Date} for computing the number of weeks
-     * @param endDate   The end {@link Date} for computing the number of weeks
-     */
     public SlidePagerAdapter(Context context, Date startDate, Date endDate) {
-        this(context, startDate, endDate, null, null, -1);
-
+        super(context, startDate, endDate);
     }
 
     public SlidePagerAdapter(Context context, Date startDate, Date endDate, TypedArray attributes, OnSlidePageChangeListener pageListener, int maxWeeks) {
-        this.mContext = context;
-        this.mStartDate = startDate;
-        this.mEndDate = endDate;
-        this.mMaxWeeks = maxWeeks;
-        this.mUserPageListener = pageListener;
-        setAttributeSet(attributes);
-        this.mViews = initViews();
-    }
-
-    @Override
-    public int getCount() {
-        return mViews.length;
-    }
-
-    @Override
-    public Object instantiateItem(ViewGroup collection, int position) {
-        SlideView currentView;
-        if (mViews.length > position) {
-            currentView = getViewForPosition(position);
-        } else {
-            return null;
-        }
-
-        collection.addView(currentView);
-        return currentView;
-    }
-
-    /**
-     * Gets the {@link SlideView} for the current pager position, uses lazy initialization
-     * to instantiate new views
-     * @param position The position in the pager
-     * @return The existing slide view, or a new one
-     */
-    private SlideView getViewForPosition(int position) {
-        if (mViews == null) {
-            this.mViews = initViews();
-        }
-
-        SlideView currentView = mViews[position];
-
-        if (currentView == null) {
-            currentView = getWeekSlide(position, mWeeks);
-            mViews[position] = currentView;
-        }
-
-        return currentView;
-    }
-
-    @Override
-    public void destroyItem(ViewGroup collection, int position, Object view) {
-        collection.removeView((View) view);
-    }
-
-    @Override
-    public boolean isViewFromObject(View view, Object object) {
-        return view == object;
-    }
-
-    @Override
-    public SlideView getCurrentView(int position) {
-        if (mViews == null || position > mViews.length - 1) {
-            return null;
-        }
-
-        return getViewForPosition(position);
+        super(context, startDate, endDate, attributes, pageListener, maxWeeks);
     }
 
     /**
@@ -187,7 +57,7 @@ public class SlidePagerAdapter extends AbstractSlidePagerAdapter<SlideView> {
      *
      * @return The number of weeks between the two dates
      */
-    private SlideView[] initViews() {
+    protected SlideView[] initViews() {
         if (mEndDate.before(mStartDate)) {
             throw new IllegalArgumentException("Start date must be before end date");
         }
@@ -203,7 +73,7 @@ public class SlidePagerAdapter extends AbstractSlidePagerAdapter<SlideView> {
      *
      * @return
      */
-    private SlideView getWeekSlide(int pagePosition, int weeks) {
+    protected SlideView getWeekSlide(int pagePosition, int weeks) {
         int weeksSince = weeks - (pagePosition + 1);
         SlideView week = new SlideView(mContext, mAttributeSet, pagePosition, mUserPageListener, getLeftText(pagePosition, mWeeks), getRightText(weeksSince));
         week.setListener(new OnSlideListener() {
@@ -233,24 +103,5 @@ public class SlidePagerAdapter extends AbstractSlidePagerAdapter<SlideView> {
             }
         });
         return week;
-    }
-
-    /**
-     * Returns the left text for the {@link SlideView}
-     *
-     * @return
-     */
-    private String getLeftText(int pagePosition, int totalWeeks) {
-
-        return Utilities.getWeekOfText(pagePosition, totalWeeks, mMaxWeeks);
-    }
-
-    /**
-     * Returns the right text for the {@link SlideView}
-     *
-     * @return
-     */
-    public String getRightText(int weeksSince) {
-        return Utilities.getWeekRangeText(mEndDate, weeksSince);
     }
 }
